Fatty Acid Composition of Various Sources

This dataset contains the fatty acid composition of various sources, including macroalgae, mangroves, saltmarsh, seagrass, plankton, and terrestrial grass. The data includes mean values and standard deviations for different fatty acids, as well as the study from which the data was sourced.

bcs_fa

Format

A data frame with 6 rows and 15 variables:

source

The source of the sample (e.g., Macroalgae, Mangrove, etc.).

Taxa

The taxa or species associated with the source.

24:0

Mean value of the fatty acid 24:0.

24:0 (SD)

Standard deviation of the fatty acid 24:0.

18:1w9

Mean value of the fatty acid 18:1w9.

18:1w9 (SD)

Standard deviation of the fatty acid 18:1w9.

18:2w6

Mean value of the fatty acid 18:2w6.

18:2w6 (SD)

Standard deviation of the fatty acid 18:2w6.

18:3w3

Mean value of the fatty acid 18:3w3.

18:3w3 (SD)

Standard deviation of the fatty acid 18:3w3.

20:4w6

Mean value of the fatty acid 20:4w6.

20:4w6 (SD)

Standard deviation of the fatty acid 20:4w6.

20:5w3

Mean value of the fatty acid 20:5w3.

20:5w3 (SD)

Standard deviation of the fatty acid 20:5w3.

Study

The study from which the data was sourced.

Source

The data and methodology are described in the following references:

Khotimchenko, S. V. (1991). Fatty acid composition of seven Sargassum species. Phytochemistry, 30(8), 2639-2641.

Meziane, T., Lee, S. Y., Mfilinge, P. L., Shin, P. K. S., Lam, M. H. W., & Tsuchiya, M. (2007). Inter-specific and geographical variations in the fatty acid composition of mangrove leaves: implications for using fatty acids as a taxonomic tool and tracers of organic matter. Marine Biology, 150, 1103-1113. DOI:10.1007/s00227-006-0424-z

Weete, J. D., Rivers, W. G., & Webert, D. J. (1970). Hydrocarbon and fatty acid distribution in the halophyte, Salicornia bigelovii. Phytochemistry, 9, 2041-2045.

Belicka, L. L., Burkholder, D., Fourqurean, J. W., Heithaus, M. R., Macko, S. A., & Jaffé, R. (2012). Stable isotope and fatty acid biomarkers of seagrass, epiphytic, and algal organic matter to consumers in a pristine seagrass ecosystem. Marine and Freshwater Research, 63(3), 1085-1097. DOI:10.1071/MF12027

Burns, K. A., Volkman, J. K., Cavanagh, J.-A., & Brinkman, D. (2003). Lipids as biomarkers for carbon cycling on the Northwest Shelf of Australia: results from a sediment trap study. Marine Chemistry, 80, 103-128.

Wang, S., Chu, T., Huang, D., Li, B., & Wu, J. (2014). Incorporation of Exotic Spartina alterniflora into Diet of Deposit-Feeding Snails in the Yangtze River Estuary Salt Marsh: Stable Isotope and Fatty Acid Analyses. Ecosystems, 17, 567-577. DOI:10.1007/s10021-013-9743-3

Examples

data(bcs_fa)
head(bcs_fa)
#>              source                                   Taxa 24:0 24:0 (SD)
#> 1        Macroalgae                              Sargassum 0.00      0.00
#> 2          Mangrove                              A. marina 2.73      1.40
#> 3          Plankton                sediment trap 200-1430m 0.00      0.00
#> 4         Saltmarsh                   Salicornia bigelovii 3.30      3.30
#> 5          Seagrass C. augustata, A antartica, H. uniervis 0.00      0.00
#> 6 Terrestrial grass                   Phragmites australis 0.29      0.09
#>   18:1w9 18:1w9 (SD) 18:2w6 18:2w6 (SD) 18:3w3 18:3w3 (SD) 20:4w6 20:4w6 (SD)
#> 1   9.46        1.38   6.04        1.87   7.17        0.96  14.27        1.97
#> 2  14.13        1.87  15.48        2.00  28.30        2.52   0.10        0.21
#> 3  18.50       10.16   0.56        0.60   1.42        0.26   0.38        0.85
#> 4   9.10        7.64  18.35        4.45  24.85       32.60   0.00        0.00
#> 5   4.00        1.00  27.00        5.00  33.00       12.00   1.00        0.50
#> 6   1.99        0.14  12.53        0.91  40.31        3.52   1.08        1.02
#>   20:5w3 20:5w3 (SD)              Study
#> 1   5.51        3.95  Khotimchenko 1991
#> 2   0.00        0.00 Meziane et al 2006
#> 3   4.68        3.97   Burns et al 2003
#> 4   0.00        0.00   Weete et al 1970
#> 5   1.00        1.00 Belicka et al 2012
#> 6   0.16        0.06    Wang et al 2014