BCS Stable Isotope Data

This dataset contains stable isotope ratios (\(\delta^{13}C\) and \(\delta^{15}N\)) and their standard deviations for various sources in the BCS study.

bcs_si

Format

A data frame with 6 rows and 6 variables:

source

The source of the sample (e.g., Mangrove, Seagrass, etc.).

d(13C/12C)

The stable isotope ratio of carbon (\(\delta^{13}C\)) in parts per thousand (‰).

d13C sd

The standard deviation of the \(\delta^{13}C\) measurements.

d(15N/14N)

The stable isotope ratio of nitrogen (\(\delta^{15}N\)) in parts per thousand (‰).

d15N sd

The standard deviation of the \(\delta^{15}N\) measurements.

Study

The study or reference from which the data were obtained.

Source

The data and methodology are described in the following references:

Nyunja, J., Ntiba, M., & Onyari, J. (2009). Carbon and nitrogen isotopic variation in mangrove ecosystems along the Kenyan coast. Estuarine, Coastal and Shelf Science, 84(3), 377-385. DOI:10.1016/j.ecss.2009.06.001

Bulmer, R. H., Lundquist, C. J., & Schwendenmann, L. (2020). Sediment carbon and nitrogen isotopes reveal variability in estuarine food web sources. Marine Ecology Progress Series, 641, 13-26. DOI:10.3354/meps13312

Goni, M. A., & Thomas, K. A. (2000). Sources and transformations of organic matter in surface soils and sediments from a Spartina alterniflora marsh. Estuarine, Coastal and Shelf Science, 50(5), 657-679. DOI:10.1006/ecss.2000.0597

Examples

data(bcs_si)
head(bcs_si)
#>              source d(13C/12C) d13C sd d(15N/14N) d15N sd
#> 1        Macroalgae      -20.4    3.10       2.40    1.20
#> 2          Mangrove      -27.7    1.40       4.40    2.60
#> 3          Plankton      -19.3    2.80       6.20    3.00
#> 4         Saltmarsh      -24.6    4.32       2.31    1.67
#> 5          Seagrass      -15.1    3.00       1.20    1.30
#> 6 Terrestrial grass      -13.3    0.50       1.69    0.50
#>                           Study
#> 1          Nyunja et al. (2009)
#> 2          Nyunja et al. (2009)
#> 3          Nyunja et al. (2009)
#> 4          Bulmer et al. (2020)
#> 5          Nyunja et al. (2009)
#> 6 Goni & Thomas (2000) Spartina