Reshape a data stream into a full source-proportion matrix

Extracts the sampled-source proportions for target from data_streams_list, in the column order given by order_ref, and appends the reconstructed unsampled-source proportion (1 - rowSums(sampled proportions)).

reshape_ref_data(data_streams_list, target = "df_stream_2", order_ref)

Arguments

data_streams_list

A named list containing df_stream_1 and df_stream_2. df_stream_1 must contain stream-1 tracer observations with one column per tracer and may contain a group column. df_stream_2 must contain one column per sampled source with values interpreted as stream-2 mixing proportions and may contain a group column. An Unsampled column is optional; if absent, the unsampled component is reconstructed as 1 - rowSums(sampled proportions).

target

The target element of data_streams_list, typically "df_stream_2".

order_ref

A character vector specifying the correct order of the source names.

Value

A matrix of N x (J + 1), with the + 1 representing the mixing proportions from the unsampled source(s).

Examples

library(MixMustR)
reshape_ref_data(synthetic_df_convergent, target = "df_stream_2",
                 order_ref = tracer_parameters$mus$source)
#>          Macroalgae     Mangrove    Saltmarsh     Seagrass         SPOM
#>   [1,] 0.0518315335 1.442806e-02 1.047310e-01 5.034907e-02 0.0008439401
#>   [2,] 0.1870115617 1.209536e-02 2.914679e-01 1.485374e-01 0.0029453966
#>   [3,] 0.2188340019 9.027669e-03 9.789854e-02 1.921419e-01 0.0022063347
#>   [4,] 0.0336145699 4.357209e-02 2.633203e-01 2.376619e-02 0.0057742109
#>   [5,] 0.0205051039 3.384030e-02 3.142780e-01 1.193179e-01 0.0016570095
#>   [6,] 0.1783547438 4.934977e-02 1.613642e-01 6.274085e-02 0.0074235635
#>   [7,] 0.0572363644 1.174537e-03 5.455797e-01 1.955273e-01 0.0027019925
#>   [8,] 0.0381165052 6.759111e-03 6.840544e-02 2.860638e-02 0.0041347579
#>   [9,] 0.0242987786 1.729988e-02 3.678296e-01 4.806615e-02 0.0075904060
#>  [10,] 0.0311612034 3.205343e-03 8.660665e-02 6.006250e-01 0.0221303453
#>  [11,] 0.0134891742 1.655176e-03 7.290602e-05 7.373032e-01 0.2301491111
#>  [12,] 0.0737696713 2.700808e-02 2.637234e-03 1.569808e-01 0.7139741299
#>  [13,] 0.0778532672 1.293331e-02 2.337686e-04 2.303645e-01 0.6688917899
#>  [14,] 0.0624319529 9.482456e-03 4.868568e-04 4.251084e-02 0.8546629348
#>  [15,] 0.3224306264 1.012730e-02 2.645585e-04 5.762650e-01 0.0824019514
#>  [16,] 0.0242891193 4.771416e-03 1.840571e-04 5.642099e-01 0.4047099492
#>  [17,] 0.0354905883 3.207264e-03 6.330287e-04 1.296709e-01 0.8298602018
#>  [18,] 0.0202890497 1.629464e-03 1.578851e-04 3.602325e-02 0.9220891733
#>  [19,] 0.0912276402 8.674778e-03 1.240103e-03 4.846581e-01 0.4064428748
#>  [20,] 0.0158919075 4.409608e-04 1.762083e-04 1.552244e-02 0.9659694734
#>  [21,] 0.6818265067 1.596736e-01 2.825858e-02 6.483920e-02 0.0021859940
#>  [22,] 0.3858218470 2.292340e-01 7.367063e-02 1.558199e-01 0.0006164244
#>  [23,] 0.0229195850 6.639484e-02 1.108054e-01 2.075414e-01 0.0033384876
#>  [24,] 0.6763836843 6.811189e-02 9.641980e-03 9.359990e-02 0.0030314068
#>  [25,] 0.1328469017 5.626358e-01 8.330971e-03 5.893355e-02 0.0007569796
#>  [26,] 0.3923087529 6.955877e-02 5.695939e-03 2.932922e-01 0.0005056108
#>  [27,] 0.3152694354 7.945385e-02 1.046299e-01 4.918239e-02 0.0024728179
#>  [28,] 0.2661471149 3.898952e-01 1.387224e-02 1.123650e-01 0.0026081503
#>  [29,] 0.1143601323 2.987777e-01 4.063202e-02 3.119139e-01 0.0011181861
#>  [30,] 0.7159984645 8.631675e-02 2.236700e-02 6.393704e-02 0.0050808706
#>  [31,] 0.8562423398 2.982556e-02 1.554435e-02 3.572218e-03 0.0077800407
#>  [32,] 0.8574765206 8.455740e-03 3.821604e-02 6.476347e-03 0.0486861404
#>  [33,] 0.4129163957 2.341296e-03 5.116947e-01 2.045337e-03 0.0232022514
#>  [34,] 0.6247516556 2.304633e-02 4.492499e-02 6.449268e-02 0.0191386332
#>  [35,] 0.6734306291 5.934129e-03 2.100132e-01 1.881580e-02 0.0663570695
#>  [36,] 0.2343323008 1.741879e-01 3.166246e-01 8.379748e-02 0.0506560572
#>  [37,] 0.3799604369 9.281678e-03 1.467692e-01 2.784971e-01 0.0268153455
#>  [38,] 0.3967780954 3.586560e-02 7.725225e-02 9.435379e-02 0.0357082312
#>  [39,] 0.4369861298 6.132618e-02 1.002187e-01 1.559312e-01 0.0253503366
#>  [40,] 0.6889028390 3.320290e-03 1.295202e-01 1.947513e-02 0.0092145032
#>  [41,] 0.0380538139 2.737893e-03 5.413750e-02 2.319273e-02 0.0151983847
#>  [42,] 0.0504543438 9.314989e-03 1.093769e-01 3.377600e-02 0.0061569798
#>  [43,] 0.3260443521 4.599309e-03 2.109755e-02 6.889635e-02 0.0214184881
#>  [44,] 0.1949529096 2.620125e-03 8.668706e-02 3.450738e-02 0.0010067225
#>  [45,] 0.4861438799 2.199983e-03 1.395901e-01 2.594268e-02 0.0229444479
#>  [46,] 0.1506007434 6.020059e-04 5.331614e-02 7.353577e-02 0.0031723779
#>  [47,] 0.0504044973 3.272062e-03 4.144544e-02 7.865938e-02 0.0297038932
#>  [48,] 0.1067081547 9.280021e-03 1.233101e-01 4.910575e-02 0.0243350741
#>  [49,] 0.0772863024 5.932565e-03 1.027628e-01 1.359206e-01 0.0059838923
#>  [50,] 0.1661279205 3.999489e-03 2.184396e-02 2.756424e-02 0.0080271516
#>  [51,] 0.0026349640 2.347879e-01 1.762681e-02 7.748629e-02 0.0660520080
#>  [52,] 0.0025247450 7.536445e-03 1.221472e-03 2.645111e-01 0.0049572137
#>  [53,] 0.0076368217 8.113785e-02 2.785525e-03 7.131590e-02 0.0980326850
#>  [54,] 0.0006675658 3.148177e-01 6.771841e-03 2.636431e-01 0.0115455910
#>  [55,] 0.0061660446 2.376856e-01 8.233702e-03 1.713522e-01 0.0286780711
#>  [56,] 0.0018490769 7.288990e-02 2.867658e-03 5.096402e-01 0.0072695920
#>  [57,] 0.0028135212 7.080783e-02 9.036972e-04 4.495944e-01 0.0248597370
#>  [58,] 0.0067887091 1.550204e-01 3.518826e-02 2.894921e-01 0.0432658639
#>  [59,] 0.0016524255 5.275825e-02 1.502800e-02 5.051417e-01 0.0178264254
#>  [60,] 0.0068688482 2.163310e-01 1.042636e-02 5.482083e-01 0.0380338594
#>  [61,] 0.0227301536 2.704601e-01 1.285018e-02 1.119591e-01 0.0003283738
#>  [62,] 0.0568790313 2.753814e-01 5.581574e-02 4.441596e-01 0.0017840149
#>  [63,] 0.1064669981 2.083930e-01 1.717416e-02 9.860182e-02 0.0012916677
#>  [64,] 0.0804091578 5.074738e-01 2.094502e-03 3.100404e-01 0.0022785552
#>  [65,] 0.0214078725 3.589373e-01 5.698335e-02 1.795950e-01 0.0011119689
#>  [66,] 0.1072928560 3.322067e-01 5.147644e-03 3.055042e-02 0.0018972513
#>  [67,] 0.3137484123 9.370035e-02 2.233827e-02 6.558421e-02 0.0026205904
#>  [68,] 0.0678414771 2.582046e-01 2.323648e-02 4.163620e-01 0.0025438315
#>  [69,] 0.1345811978 3.687180e-01 1.007758e-02 9.694359e-02 0.0011086777
#>  [70,] 0.0548517897 3.012471e-02 1.192553e-02 5.767518e-02 0.0004518031
#>  [71,] 0.0715846397 1.294182e-02 1.685047e-01 5.512130e-01 0.0031015613
#>  [72,] 0.0393789615 1.160179e-02 8.783437e-02 3.633949e-01 0.0069668694
#>  [73,] 0.0020237595 4.626658e-03 3.965353e-02 2.968559e-01 0.0018329632
#>  [74,] 0.0080546821 1.369060e-03 1.305887e-01 8.099432e-01 0.0018777145
#>  [75,] 0.0051694450 2.783242e-02 4.461315e-02 9.818329e-02 0.0082457066
#>  [76,] 0.0025805649 1.664050e-02 1.472618e-01 6.022399e-01 0.0027721093
#>  [77,] 0.0022947741 2.687864e-02 5.310876e-02 7.448423e-01 0.0002726908
#>  [78,] 0.0121755319 7.500220e-02 1.471414e-01 5.260994e-01 0.0134880211
#>  [79,] 0.0080631835 6.749653e-03 1.163867e-02 3.074509e-01 0.0051895306
#>  [80,] 0.0187138438 1.696894e-01 4.788231e-02 4.197113e-01 0.0028411764
#>  [81,] 0.0110986345 7.860324e-01 2.522164e-02 7.074643e-05 0.0448140397
#>  [82,] 0.0098560604 8.670648e-01 6.280116e-03 1.266867e-04 0.0480290092
#>  [83,] 0.0008479802 8.212812e-01 1.415050e-02 8.307674e-05 0.0561922227
#>  [84,] 0.1552814259 5.675335e-01 9.392234e-02 6.335578e-05 0.0823036684
#>  [85,] 0.0081746708 9.229053e-01 8.744469e-04 1.941200e-05 0.0619808625
#>  [86,] 0.0110502467 9.527890e-01 2.930778e-03 1.642100e-05 0.0178526692
#>  [87,] 0.0082800916 5.023756e-01 5.580464e-03 7.561864e-06 0.4444547665
#>  [88,] 0.0059091721 9.054768e-01 1.906630e-02 7.077856e-05 0.0566903996
#>  [89,] 0.0046435833 8.403989e-01 8.069837e-02 5.209621e-05 0.0073079341
#>  [90,] 0.0057288724 9.655149e-01 1.361172e-02 5.204417e-06 0.0089553863
#>  [91,] 0.0417829101 2.966084e-04 3.496857e-04 1.888779e-02 0.9289752479
#>  [92,] 0.0007416609 3.152308e-05 9.162705e-05 2.810971e-03 0.9780704043
#>  [93,] 0.1068574630 4.280822e-04 3.486056e-03 8.838684e-02 0.7580902739
#>  [94,] 0.0005765537 1.669791e-05 6.376185e-05 1.357261e-03 0.9901394944
#>  [95,] 0.0021367236 4.280148e-04 5.775549e-04 7.788732e-03 0.9745503631
#>  [96,] 0.0256401948 4.956791e-04 5.153279e-04 3.144507e-02 0.9306252868
#>  [97,] 0.0098006307 9.239731e-04 1.467314e-04 1.411702e-01 0.5698288479
#>  [98,] 0.0044652949 4.651544e-04 8.115735e-04 8.756038e-03 0.9840861062
#>  [99,] 0.0019038907 7.086415e-05 6.335694e-05 2.007096e-01 0.7653903260
#> [100,] 0.0464229878 8.871414e-04 8.767910e-04 8.327400e-03 0.8774169587
#>        Terrestrial grass    Unsampled
#>   [1,]      0.3171890764 0.4606272874
#>   [2,]      0.1588287176 0.1991136110
#>   [3,]      0.2566249074 0.2232666740
#>   [4,]      0.0244343370 0.6055182561
#>   [5,]      0.0324261813 0.4779755689
#>   [6,]      0.2244751792 0.3162916502
#>   [7,]      0.0050967219 0.1926833890
#>   [8,]      0.0084587726 0.8455190360
#>   [9,]      0.0322104313 0.5027047130
#>  [10,]      0.0536873952 0.2025840551
#>  [11,]      0.0020940127 0.0152363788
#>  [12,]      0.0094955145 0.0161345299
#>  [13,]      0.0006548317 0.0090685540
#>  [14,]      0.0007809462 0.0296440149
#>  [15,]      0.0016961621 0.0068143748
#>  [16,]      0.0005898039 0.0012457661
#>  [17,]      0.0004100383 0.0007280191
#>  [18,]      0.0001616415 0.0196495381
#>  [19,]      0.0028963683 0.0048601664
#>  [20,]      0.0012755190 0.0007234949
#>  [21,]      0.0376076784 0.0256084317
#>  [22,]      0.1164217092 0.0384154388
#>  [23,]      0.4485332428 0.1404670657
#>  [24,]      0.1447280331 0.0045031020
#>  [25,]      0.2327212096 0.0037746100
#>  [26,]      0.0419603664 0.1966783992
#>  [27,]      0.4412951958 0.0076963692
#>  [28,]      0.2064068395 0.0087054088
#>  [29,]      0.2166630323 0.0165351002
#>  [30,]      0.0921686765 0.0141312039
#>  [31,]      0.0669642640 0.0200712274
#>  [32,]      0.0132910752 0.0273981339
#>  [33,]      0.0457220218 0.0020779690
#>  [34,]      0.2020491937 0.0215965192
#>  [35,]      0.0233199434 0.0021292090
#>  [36,]      0.0942014713 0.0462001766
#>  [37,]      0.0575776967 0.1010985226
#>  [38,]      0.2999144200 0.0601276226
#>  [39,]      0.2131659439 0.0070215406
#>  [40,]      0.0737544100 0.0758126764
#>  [41,]      0.7997502890 0.0669293853
#>  [42,]      0.6864543142 0.1044664886
#>  [43,]      0.4042638721 0.1536800709
#>  [44,]      0.3744310877 0.3057947178
#>  [45,]      0.2255139509 0.0976649670
#>  [46,]      0.3082582369 0.4105147239
#>  [47,]      0.7471242501 0.0493904727
#>  [48,]      0.5755314951 0.1117293535
#>  [49,]      0.4357482691 0.2363656631
#>  [50,]      0.6272737242 0.1451635157
#>  [51,]      0.4888007190 0.1126112884
#>  [52,]      0.5522889173 0.1669601382
#>  [53,]      0.6171386437 0.1219525673
#>  [54,]      0.3473760239 0.0551781897
#>  [55,]      0.2720103309 0.2758740223
#>  [56,]      0.1671298233 0.2383537415
#>  [57,]      0.3014075070 0.1496133301
#>  [58,]      0.1299147672 0.3403298620
#>  [59,]      0.3919535527 0.0156396693
#>  [60,]      0.1468032160 0.0333283814
#>  [61,]      0.0482038486 0.5334681457
#>  [62,]      0.0818415472 0.0841386504
#>  [63,]      0.0175617809 0.5505105552
#>  [64,]      0.0262095341 0.0714941032
#>  [65,]      0.1007883699 0.2811761837
#>  [66,]      0.1433953384 0.3795098335
#>  [67,]      0.0864672525 0.4155409199
#>  [68,]      0.0917907460 0.1400208460
#>  [69,]      0.1151062761 0.2734646364
#>  [70,]      0.1558355167 0.6891354727
#>  [71,]      0.0413172542 0.1513369593
#>  [72,]      0.0470980811 0.4437249954
#>  [73,]      0.5214164853 0.1335907219
#>  [74,]      0.0422285446 0.0059380913
#>  [75,]      0.1763777224 0.6395782592
#>  [76,]      0.2139114504 0.0145936572
#>  [77,]      0.1271728548 0.0454299624
#>  [78,]      0.1431760486 0.0829173364
#>  [79,]      0.3928644957 0.2680435475
#>  [80,]      0.2318227081 0.1093393233
#>  [81,]      0.0139256847 0.1188368267
#>  [82,]      0.0095585136 0.0590847895
#>  [83,]      0.0058162706 0.1016287469
#>  [84,]      0.0836868157 0.0172088839
#>  [85,]      0.0039401787 0.0021051374
#>  [86,]      0.0020379442 0.0133229296
#>  [87,]      0.0046681141 0.0346334385
#>  [88,]      0.0061835443 0.0066030459
#>  [89,]      0.0136343456 0.0532647438
#>  [90,]      0.0028944785 0.0032894293
#>  [91,]      0.0081240896 0.0015836655
#>  [92,]      0.0181029908 0.0001508226
#>  [93,]      0.0198951204 0.0228561630
#>  [94,]      0.0044253620 0.0034208696
#>  [95,]      0.0132669586 0.0012516533
#>  [96,]      0.0091256428 0.0021528002
#>  [97,]      0.2759831700 0.0021463987
#>  [98,]      0.0005226303 0.0008932025
#>  [99,]      0.0313113071 0.0005506254
#> [100,]      0.0589139498 0.0071547713