run_mixmustr_models

Runs a set of Stan models.

run_mixmustr_models(
  model_choices,
  data_streams_list,
  mu_tab,
  sig_tab = NULL,
  sigma_ln_rho,
  ...
)

Arguments

model_choices

A data frame specifying the model configurations to run. Each row should represent a model, with columns indicating the values for sample_tracer, fix_unsampled, hierarchical, and code_path.

data_streams_list

A list containing the input data streams for the model. It should include two data frames named df_stream_1 anddf_stream_2. See details for the expected structure of these.

mu_tab

A data frame containing the mean tracer signatures for each source. The first column should be named source, and the remaining columns should contain numeric tracer values.

sig_tab

(Optional) A data frame containing the standard deviations of tracer signatures for each source. The structure must match mu_tab. Required if sample_tracer is TRUE or if fix_unsampled is FALSE and sample_tracer is FALSE.

sigma_ln_rho

A numeric value or matrix specifying the confidence around the log mixing proportions. If a matrix, it must be of dimensions N X J, with N being the number of observations and J being the number of sources.

...

Additional arguments passed to stan, such as iter, chains, or control.

Value

A list where each element corresponds to a model configuration, containing the run time and model output.

Details

The run_mixmustr_models function automates the process of running multiple mixmustr models based on user-specified configurations. It builds the Stan code for each model, prepares the input data, and runs the models sequentially. The results are returned as a list, with each element containing the run time and model output for a specific configuration. For convenience, mixmustr provides the user with a built-in data frame containing all of the potential allowed models mixmustr_models.

See also

Examples

if (FALSE) { # \dontrun{
library(mixmustr)
mus <- tracer_parameters$mus
sigmas <- tracer_parameters$sigmas
model_fits <- run_mixmustr_models(
  mixmustr_models[1, ], synthetic_df_convergent, mus, sigmas,
  sigma_ln_rho = 0.1, iter = 1e4, warmup = 5e3, chains = 4, cores = 4
)
} # }