Synthetic Convergent Dataset

This dataset is a synthetic dataset generated to simulate mixture data for testing and validation purposes in the MixMustR package. It is anchored to empirical values of stable isotopes and fatty acids for a range of plant carbon sources in marine soils stable isotope data (bcs_si), fatty acid data (bcs_fa), and synthetic mixture proportions. synthetic_df_convergent exhibits little difference in the underlying mixing proportions between data streams 1 and 2.

synthetic_df_convergent

Format

A list containing three data frames:

df_stream_1

A data frame containing the simulated mixture data for the first stream, including tracer estimates calculated from the stable isotope and fatty acid data.

df_stream_2

A data frame containing the simulated synthetic proportions for the second stream, a column for each source.

stream_1_props

A data frame containing the synthetic proportions for the first stream, used as input for generating the mixture data. This is only used for testing purposes

Source

The synthetic dataset was generated programmatically using the MixMustR package, following the simulation design described in the accompanying manuscript. The input data sources are:

Stable isotope data: See bcs_si documentation.

Fatty acid data: See bcs_fa documentation.

Pooled-unsampled tracer signature: synthetically constructed, offset from the sampled-source centroid (see make_unsampled_signature).

Details

The non-exported augment_bcs_with_unsampled function first combines stable isotope data (bcs_si) and fatty acid data (bcs_fa), homogenises their sampled-source SD/sample-size columns, and appends a synthetic pooled-unsampled tracer signature (built by the non-exported make_unsampled_signature function) positioned away from the sampled-source multivariate space. The non-exported simulate_mvn_mixture function then generates the two composition streams (stream_1_props and df_stream_2) using a hierarchical logistic-normal process across 10 groups of 10 observations. The non-exported make_mixture_data function combines these signatures and proportions to produce the simulated tracer measurements in df_stream_1; the truth_stream parameter determines which stream's proportions are used as the "true" source contributions. Finally, the non-exported rm_unsampled function removes the explicit Unsampled proportion column from df_stream_2 and stream_1_props, mimicking the realistic situation in which users do not supply it directly; mixmustr_wrangle_input reconstructs it internally.

See also

Examples

library(MixMustR)
data(synthetic_df_convergent)
str(synthetic_df_convergent)
#> List of 3
#>  $ df_stream_1   :'data.frame':	100 obs. of  9 variables:
#>   ..$ group     : chr [1:100] "G1" "G1" "G1" "G1" ...
#>   ..$ 18:1w9    : num [1:100] 11.84 10.32 9.81 14.11 12.83 ...
#>   ..$ 18:2w6    : num [1:100] 7.38 11.27 10.26 6.73 9.37 ...
#>   ..$ 18:3w3    : num [1:100] 15.3 18.2 18.3 10.3 13.6 ...
#>   ..$ 20:4w6    : num [1:100] 5.01 5.14 5.54 5.37 4.73 ...
#>   ..$ 20:5w3    : num [1:100] 2.93 2.63 2.78 3.3 2.89 ...
#>   ..$ 24:0      : num [1:100] 1.67 1.8 1.5 1.92 1.93 ...
#>   ..$ d(13C/12C): num [1:100] -19 -19.7 -18.6 -21.2 -20.8 ...
#>   ..$ d(15N/14N): num [1:100] 3.39 2.7 2.91 3.6 3.25 ...
#>  $ df_stream_2   :'data.frame':	100 obs. of  7 variables:
#>   ..$ group            : chr [1:100] "G1" "G1" "G1" "G1" ...
#>   ..$ Macroalgae       : num [1:100] 0.0518 0.187 0.2188 0.0336 0.0205 ...
#>   ..$ Mangrove         : num [1:100] 0.01443 0.0121 0.00903 0.04357 0.03384 ...
#>   ..$ Saltmarsh        : num [1:100] 0.1047 0.2915 0.0979 0.2633 0.3143 ...
#>   ..$ Seagrass         : num [1:100] 0.0503 0.1485 0.1921 0.0238 0.1193 ...
#>   ..$ SPOM             : num [1:100] 0.000844 0.002945 0.002206 0.005774 0.001657 ...
#>   ..$ Terrestrial grass: num [1:100] 0.3172 0.1588 0.2566 0.0244 0.0324 ...
#>  $ stream_1_props:'data.frame':	100 obs. of  7 variables:
#>   ..$ group            : chr [1:100] "G1" "G1" "G1" "G1" ...
#>   ..$ Macroalgae       : num [1:100] 0.0619 0.1533 0.1773 0.044 0.0309 ...
#>   ..$ Mangrove         : num [1:100] 0.0157 0.0139 0.0117 0.0331 0.0276 ...
#>   ..$ Saltmarsh        : num [1:100] 0.137 0.282 0.135 0.254 0.286 ...
#>   ..$ Seagrass         : num [1:100] 0.0633 0.1359 0.1687 0.036 0.112 ...
#>   ..$ SPOM             : num [1:100] 0.00143 0.00345 0.00291 0.00537 0.00221 ...
#>   ..$ Terrestrial grass: num [1:100] 0.2139 0.1311 0.1904 0.0337 0.041 ...