Tracer Parameters

This built-in list contains standardised mean (\(\mu\)), standard deviation (\(\sigma\)) and sample size values for stable isotope and fatty acid tracers, derived from the bcs_si and bcs_fa datasets. It is used internally by the MixMustR package for modelling and analysis.

tracer_parameters

Format

A list with three elements:

mus

A data frame containing the mean (\(\mu\)) values for each tracer, grouped by source. Columns represent tracers, and rows represent sources.

sigmas

A data frame containing the standard deviation (\(\sigma\)) values for each tracer, grouped by source. Columns represent tracers, and rows represent sources.

ns

A data frame containing the sample size values for each tracer, grouped by source. Columns represent tracers, and rows represent sources.

Source

The data is derived from the bcs_si and bcs_fa datasets. See their respective documentation for more details.

Details

The tracer_parameters list is generated using the MixMustR:::wrangle_tracer_pars function, which processes the bcs_si and bcs_fa datasets to reshape and combine their tracer data. The mus element contains the mean values for each tracer, the sigmas element contains the corresponding standard deviations and the ns element contains the sample sizes. These values are used as inputs for mixture modelling in the MixMustR package.

Examples

library(MixMustR)
data(tracer_parameters)
str(tracer_parameters)
#> List of 3
#>  $ mus   : tibble [6 × 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ source    : chr [1:6] "Macroalgae" "Mangrove" "Saltmarsh" "Seagrass" ...
#>   ..$ d(13C/12C): num [1:6] -20.4 -27.7 -24.6 -15.1 -19.3 -13.3
#>   ..$ d(15N/14N): num [1:6] 2.4 4.4 2.31 1.2 6.2 1.69
#>   ..$ 24:0      : num [1:6] 0 2.73 3.3 0 0 0.29
#>   ..$ 18:1w9    : num [1:6] 9.46 14.13 9.1 4 18.5 ...
#>   ..$ 18:2w6    : num [1:6] 6.04 15.48 18.35 27 0.56 ...
#>   ..$ 18:3w3    : num [1:6] 7.17 28.3 24.85 33 1.42 ...
#>   ..$ 20:4w6    : num [1:6] 14.271 0.105 0 1 0.38 ...
#>   ..$ 20:5w3    : num [1:6] 5.51 0 0 1 4.68 ...
#>  $ sigmas: tibble [6 × 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ source    : chr [1:6] "Macroalgae" "Mangrove" "Saltmarsh" "Seagrass" ...
#>   ..$ d(13C/12C): num [1:6] 3.1 1.4 4.32 3 2.8 0.5
#>   ..$ d(15N/14N): num [1:6] 1.2 2.6 1.67 1.3 3 0.5
#>   ..$ 24:0      : num [1:6] 0.01 1.4 3.3 0.01 0.01 ...
#>   ..$ 18:1w9    : num [1:6] 1.38 1.87 7.64 1 10.16 ...
#>   ..$ 18:2w6    : num [1:6] 1.871 2.002 4.455 5 0.602 ...
#>   ..$ 18:3w3    : num [1:6] 0.955 2.523 32.598 12 0.259 ...
#>   ..$ 20:4w6    : num [1:6] 1.97 0.21 0.01 0.5 0.85 ...
#>   ..$ 20:5w3    : num [1:6] 3.95 0.01 0.01 1 3.97 ...
#>  $ ns    : tibble [6 × 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ source    : chr [1:6] "Macroalgae" "Mangrove" "Saltmarsh" "Seagrass" ...
#>   ..$ d(13C/12C): int [1:6] 12 15 13 20 24 15
#>   ..$ d(15N/14N): int [1:6] 12 15 13 20 24 15
#>   ..$ 24:0      : int [1:6] 10 10 10 10 10 10
#>   ..$ 18:1w9    : int [1:6] 10 10 10 10 10 10
#>   ..$ 18:2w6    : int [1:6] 10 10 10 10 10 10
#>   ..$ 18:3w3    : int [1:6] 10 10 10 10 10 10
#>   ..$ 20:4w6    : int [1:6] 10 10 10 10 10 10
#>   ..$ 20:5w3    : int [1:6] 10 10 10 10 10 10

# Access mean values
tracer_parameters$mus
#> # A tibble: 6 × 9
#>   source    `d(13C/12C)` `d(15N/14N)` `24:0` `18:1w9` `18:2w6` `18:3w3` `20:4w6`
#>   <chr>            <dbl>        <dbl>  <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1 Macroalg…        -20.4         2.4    0        9.46     6.04     7.17   14.3  
#> 2 Mangrove         -27.7         4.4    2.73    14.1     15.5     28.3     0.105
#> 3 Saltmarsh        -24.6         2.31   3.3      9.1     18.4     24.8     0    
#> 4 Seagrass         -15.1         1.2    0        4       27       33       1    
#> 5 SPOM             -19.3         6.2    0       18.5      0.56     1.42    0.38 
#> 6 Terrestr…        -13.3         1.69   0.29     1.99    12.5     40.3     1.08 
#> # ℹ 1 more variable: `20:5w3` <dbl>

# Access standard deviation values
tracer_parameters$sigmas
#> # A tibble: 6 × 9
#>   source    `d(13C/12C)` `d(15N/14N)` `24:0` `18:1w9` `18:2w6` `18:3w3` `20:4w6`
#>   <chr>            <dbl>        <dbl>  <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1 Macroalg…         3.1          1.2    0.01     1.38    1.87     0.955    1.97 
#> 2 Mangrove          1.4          2.6    1.40     1.87    2.00     2.52     0.21 
#> 3 Saltmarsh         4.32         1.67   3.3      7.64    4.45    32.6      0.01 
#> 4 Seagrass          3            1.3    0.01     1       5       12        0.5  
#> 5 SPOM              2.8          3      0.01    10.2     0.602    0.259    0.850
#> 6 Terrestr…         0.5          0.5    0.09     0.14    0.91     3.52     1.02 
#> # ℹ 1 more variable: `20:5w3` <dbl>

# Access sample sizes
tracer_parameters$ns
#> # A tibble: 6 × 9
#>   source    `d(13C/12C)` `d(15N/14N)` `24:0` `18:1w9` `18:2w6` `18:3w3` `20:4w6`
#>   <chr>            <int>        <int>  <int>    <int>    <int>    <int>    <int>
#> 1 Macroalg…           12           12     10       10       10       10       10
#> 2 Mangrove            15           15     10       10       10       10       10
#> 3 Saltmarsh           13           13     10       10       10       10       10
#> 4 Seagrass            20           20     10       10       10       10       10
#> 5 SPOM                24           24     10       10       10       10       10
#> 6 Terrestr…           15           15     10       10       10       10       10
#> # ℹ 1 more variable: `20:5w3` <int>