Fatty Acid Composition of Various Sources

This dataset contains the fatty acid composition of various sources, including macroalgae, mangroves, saltmarsh, seagrass, SPOM (suspended particulate organic matter), and terrestrial grass. The data includes mean values, standard deviations and sample sizes for different fatty acids, as well as the study from which the data was sourced. Values of 0 standard deviation have been arbitrarily transformed to 0.01.

bcs_fa

Format

A data frame with 6 rows and 15 variables:

source

The source of the sample (e.g., Macroalgae, Mangrove, etc.).

Taxa

The taxa or species associated with the source.

24:0

Mean value of the fatty acid 24:0.

24:0 (SD)

Standard deviation of the fatty acid 24:0.

24:0 (n)

Sample size of the fatty acid 24:0.

18:1w9

Mean value of the fatty acid 18:1w9.

18:1w9 (SD)

Standard deviation of the fatty acid 18:1w9.

18:1w9 (n)

Sample size of the fatty acid 18:1w9.

18:2w6

Mean value of the fatty acid 18:2w6.

18:2w6 (SD)

Standard deviation of the fatty acid 18:2w6.

18:2w6 (n)

Sample size of the fatty acid 18:2w6.

18:3w3

Mean value of the fatty acid 18:3w3.

18:3w3 (SD)

Standard deviation of the fatty acid 18:3w3.

18:3w3 (n)

Sample size of the fatty acid 18:3w3.

20:4w6

Mean value of the fatty acid 20:4w6.

20:4w6 (SD)

Standard deviation of the fatty acid 20:4w6.

20:4w6 (n)

Sample size of the fatty acid 20:4w6.

20:5w3

Mean value of the fatty acid 20:5w3.

20:5w3 (SD)

Standard deviation of the fatty acid 20:5w3.

20:5w3 (n)

Sample size of the fatty acid 20:5w3.

Study

The study from which the data was sourced.

Source

The data and methodology are described in the following references:

Khotimchenko, S. V. (1991). Fatty acid composition of seven Sargassum species. Phytochemistry, 30(8), 2639-2641.

Meziane, T., Lee, S. Y., Mfilinge, P. L., Shin, P. K. S., Lam, M. H. W., & Tsuchiya, M. (2007). Inter-specific and geographical variations in the fatty acid composition of mangrove leaves: implications for using fatty acids as a taxonomic tool and tracers of organic matter. Marine Biology, 150, 1103-1113. DOI:10.1007/s00227-006-0424-z

Weete, J. D., Rivers, W. G., & Webert, D. J. (1970). Hydrocarbon and fatty acid distribution in the halophyte, Salicornia bigelovii. Phytochemistry, 9, 2041-2045.

Belicka, L. L., Burkholder, D., Fourqurean, J. W., Heithaus, M. R., Macko, S. A., & Jaffé, R. (2012). Stable isotope and fatty acid biomarkers of seagrass, epiphytic, and algal organic matter to consumers in a pristine seagrass ecosystem. Marine and Freshwater Research, 63(3), 1085-1097. DOI:10.1071/MF12027

Burns, K. A., Volkman, J. K., Cavanagh, J.-A., & Brinkman, D. (2003). Lipids as biomarkers for carbon cycling on the Northwest Shelf of Australia: results from a sediment trap study. Marine Chemistry, 80, 103-128.

Wang, S., Chu, T., Huang, D., Li, B., & Wu, J. (2014). Incorporation of Exotic Spartina alterniflora into Diet of Deposit-Feeding Snails in the Yangtze River Estuary Salt Marsh: Stable Isotope and Fatty Acid Analyses. Ecosystems, 17, 567-577. DOI:10.1007/s10021-013-9743-3

Examples

data(bcs_fa)
head(bcs_fa)
#>              source                                   Taxa 24:0 24:0 (SD)
#> 1        Macroalgae                              Sargassum 0.00  0.010000
#> 2          Mangrove                              A. marina 2.73  1.398428
#> 3              SPOM                sediment trap 200-1430m 0.00  0.010000
#> 4         Saltmarsh                   Salicornia bigelovii 3.30  3.300000
#> 5          Seagrass C. augustata, A antartica, H. uniervis 0.00  0.010000
#> 6 Terrestrial grass                   Phragmites australis 0.29  0.090000
#>   24:0 (n)    18:1w9 18:1w9 (SD) 18:1w9 (n)    18:2w6 18:2w6 (SD) 18:2w6 (n)
#> 1       10  9.457143    1.384265         10  6.042857   1.8707014         10
#> 2       10 14.127500    1.867822         10 15.480000   2.0021988         10
#> 3       10 18.500000   10.163415         10  0.560000   0.6024948         10
#> 4       10  9.100000    7.636753         10 18.350000   4.4547727         10
#> 5       10  4.000000    1.000000         10 27.000000   5.0000000         10
#> 6       10  1.990000    0.140000         10 12.530000   0.9100000         10
#>      18:3w3 18:3w3 (SD) 18:3w3 (n)   20:4w6 20:4w6 (SD) 20:4w6 (n)   20:5w3
#> 1  7.171429   0.9551863         10 14.27143   1.9652941         10 5.514286
#> 2 28.295000   2.5227564         10  0.10500   0.2100000         10 0.000000
#> 3  1.420000   0.2588436         10  0.38000   0.8497058         10 4.680000
#> 4 24.850000  32.5976226         10  0.00000   0.0100000         10 0.000000
#> 5 33.000000  12.0000000         10  1.00000   0.5000000         10 1.000000
#> 6 40.310000   3.5200000         10  1.08000   1.0200000         10 0.160000
#>   20:5w3 (SD) 20:5w3 (n)              Study
#> 1    3.948176         10  Khotimchenko 1991
#> 2    0.010000         10 Meziane et al 2006
#> 3    3.965728         10   Burns et al 2003
#> 4    0.010000         10   Weete et al 1970
#> 5    1.000000         10 Belicka et al 2012
#> 6    0.060000         10    Wang et al 2014