Calculates unsampled mixing proportions

Calculates unsampled mixing proportions

reshape_ref_data(data_streams_list, target = "df_stream_2", order_ref)

Arguments

data_streams_list

A list containing the input data streams for the model. It should include two data frames named df_stream_1 anddf_stream_2. See details for the expected structure of these.

target

The target element of data_streams_list, typically "df_stream_2".

order_ref

A character vector specifying the correct order of the source names.

Value

A matrix of N x (J + 1), with the + 1 representing the mixing proportions from the unsampled source(s).

Examples

library(MixMustR)
reshape_ref_data(synthetic_df_convergent, target = "df_stream_2",
                 order_ref = tracer_parameters$mus$source)
#>        Macroalgae    Mangrove  Saltmarsh   Seagrass        SPOM
#>  [1,] 0.000000000 0.426265543 0.11692895 0.30083088 0.039650024
#>  [2,] 0.070792164 0.247829657 0.32877222 0.20512423 0.069206431
#>  [3,] 0.040246162 0.250371487 0.19613769 0.28995024 0.115196877
#>  [4,] 0.049396259 0.016378917 0.36233438 0.41302643 0.032413303
#>  [5,] 0.027984711 0.232983182 0.36900755 0.27004438 0.000000000
#>  [6,] 0.135730204 0.318667988 0.22675695 0.16011274 0.034515139
#>  [7,] 0.076314383 0.193230352 0.26106294 0.21981257 0.000000000
#>  [8,] 0.045254593 0.278489535 0.24981926 0.15427458 0.000000000
#>  [9,] 0.000000000 0.222982633 0.16758899 0.35145409 0.046269316
#> [10,] 0.096381480 0.191004216 0.39477933 0.24575298 0.000000000
#> [11,] 0.050523547 0.445951179 0.05151986 0.15160935 0.000000000
#> [12,] 0.151547641 0.173920201 0.09463927 0.35746474 0.148353795
#> [13,] 0.196935107 0.471211595 0.02929100 0.20363364 0.079458203
#> [14,] 0.154529178 0.193580584 0.32327176 0.11100173 0.000000000
#> [15,] 0.097099029 0.259274871 0.17197046 0.27911871 0.100706058
#> [16,] 0.010804465 0.204172401 0.24924046 0.19310614 0.107518780
#> [17,] 0.183663635 0.158134131 0.47358516 0.07896367 0.000000000
#> [18,] 0.000000000 0.408788254 0.32816414 0.06544404 0.102042625
#> [19,] 0.000000000 0.336908567 0.26974726 0.13593743 0.071140519
#> [20,] 0.164192229 0.222164431 0.31019089 0.08642454 0.132102497
#> [21,] 0.000000000 0.285597292 0.33979685 0.16834563 0.104085025
#> [22,] 0.100667525 0.197948803 0.24218826 0.15162227 0.208048299
#> [23,] 0.006079658 0.259783591 0.45644861 0.00000000 0.087608455
#> [24,] 0.094226439 0.231654151 0.34203656 0.12188233 0.082411454
#> [25,] 0.053661972 0.137667451 0.19437840 0.13656983 0.360885153
#> [26,] 0.007572000 0.220507178 0.27483082 0.25772513 0.112943645
#> [27,] 0.117088836 0.255371984 0.29561612 0.10051613 0.105370333
#> [28,] 0.177175829 0.221511893 0.25515942 0.27154944 0.036375355
#> [29,] 0.085848164 0.258576244 0.18810875 0.28816745 0.047842688
#> [30,] 0.000000000 0.208304972 0.23917330 0.10132881 0.116414101
#> [31,] 0.000000000 0.068799774 0.42974664 0.23961803 0.041920222
#> [32,] 0.099323307 0.059503656 0.50476081 0.14608496 0.000000000
#> [33,] 0.192900361 0.242891031 0.41636740 0.04370208 0.027959054
#> [34,] 0.166598452 0.209604520 0.13218459 0.22217239 0.117392454
#> [35,] 0.000000000 0.414411427 0.17762783 0.11118664 0.070382655
#> [36,] 0.180028176 0.691688513 0.00000000 0.01651992 0.008299435
#> [37,] 0.211689786 0.318424650 0.19055523 0.12093133 0.057386363
#> [38,] 0.042478317 0.000000000 0.68702156 0.09891348 0.024124817
#> [39,] 0.140028723 0.090770604 0.42711373 0.11886931 0.085425015
#> [40,] 0.214243652 0.128436604 0.20745952 0.14635290 0.103726321
#> [41,] 0.034360596 0.000000000 0.14192921 0.34327197 0.052576049
#> [42,] 0.090108106 0.037126967 0.25139264 0.26321022 0.125352345
#> [43,] 0.169405688 0.102695976 0.17414585 0.26852120 0.074033836
#> [44,] 0.201829250 0.006426153 0.06094002 0.55715056 0.117819335
#> [45,] 0.148219197 0.000000000 0.18704570 0.30103500 0.156958617
#> [46,] 0.160658762 0.165762617 0.06807909 0.23323986 0.038600775
#> [47,] 0.064705771 0.061512498 0.03592176 0.52713997 0.000000000
#> [48,] 0.000000000 0.259902389 0.03518434 0.28197764 0.128314637
#> [49,] 0.111565955 0.039952258 0.06957165 0.35395983 0.104832614
#> [50,] 0.090597890 0.000000000 0.06027805 0.27450684 0.092002977
#> [51,] 0.000000000 0.019889423 0.10987836 0.18084660 0.111673800
#> [52,] 0.144105165 0.071170414 0.19980570 0.32909181 0.061242419
#> [53,] 0.104129128 0.126196524 0.09519173 0.19981385 0.090279007
#> [54,] 0.153260058 0.132833130 0.13861447 0.28547185 0.085405947
#> [55,] 0.113663601 0.112439973 0.09902472 0.43020718 0.000000000
#> [56,] 0.023451619 0.000000000 0.30071624 0.29324828 0.094621179
#> [57,] 0.154137042 0.210574513 0.10462508 0.16370143 0.040952724
#> [58,] 0.073504528 0.129359343 0.10364724 0.41719127 0.000000000
#> [59,] 0.097718821 0.101005883 0.09458351 0.27801957 0.089823160
#> [60,] 0.083091979 0.105497432 0.11331939 0.28372989 0.032082483
#> [61,] 0.000000000 0.104457842 0.10085556 0.33175833 0.071625301
#> [62,] 0.128405154 0.123563660 0.11771390 0.36458490 0.108761272
#> [63,] 0.132252791 0.030755778 0.00000000 0.44530856 0.094259226
#> [64,] 0.220665299 0.153087619 0.10488329 0.23315906 0.150229617
#> [65,] 0.079030725 0.155537041 0.02395136 0.38479643 0.125311156
#> [66,] 0.278981427 0.117129017 0.12599321 0.31201943 0.046309309
#> [67,] 0.343583966 0.000000000 0.01721511 0.41340666 0.051088734
#> [68,] 0.045953583 0.107354515 0.11070104 0.21472600 0.000000000
#> [69,] 0.021550646 0.289500393 0.00000000 0.44821257 0.022053542
#> [70,] 0.109489121 0.071220121 0.08577507 0.35361788 0.036078418
#> [71,] 0.184240733 0.093368368 0.09923697 0.41005284 0.064359435
#> [72,] 0.000000000 0.210325886 0.04874800 0.37272279 0.060603019
#> [73,] 0.165030217 0.224193942 0.12251725 0.32201486 0.060914639
#> [74,] 0.164156050 0.012440014 0.08259267 0.27296802 0.000000000
#> [75,] 0.000000000 0.200757207 0.08587388 0.46841793 0.036641358
#> [76,] 0.062372360 0.130675721 0.12690085 0.29584021 0.029135112
#> [77,] 0.085803991 0.089231697 0.11996747 0.34668972 0.132183639
#> [78,] 0.103649428 0.040122225 0.00000000 0.42658966 0.098183303
#> [79,] 0.173040868 0.036765471 0.05730048 0.27940164 0.000000000
#>       Terrestrial grass    Unsampled
#>  [1,]       0.078964451 3.736015e-02
#>  [2,]       0.078275293 0.000000e+00
#>  [3,]       0.108097544 0.000000e+00
#>  [4,]       0.000000000 1.264507e-01
#>  [5,]       0.076994641 2.298553e-02
#>  [6,]       0.124216986 1.110223e-16
#>  [7,]       0.155562555 9.401720e-02
#>  [8,]       0.086905413 1.852566e-01
#>  [9,]       0.052408149 1.592968e-01
#> [10,]       0.069079463 3.002538e-03
#> [11,]       0.182219466 1.181766e-01
#> [12,]       0.074074358 1.110223e-16
#> [13,]       0.019470448 2.220446e-16
#> [14,]       0.112358151 1.052586e-01
#> [15,]       0.091830873 0.000000e+00
#> [16,]       0.000000000 2.351577e-01
#> [17,]       0.003688429 1.019650e-01
#> [18,]       0.066816363 2.874457e-02
#> [19,]       0.126639707 5.962651e-02
#> [20,]       0.084925418 1.110223e-16
#> [21,]       0.012692666 8.948253e-02
#> [22,]       0.099524840 0.000000e+00
#> [23,]       0.002825795 1.872539e-01
#> [24,]       0.127789057 1.110223e-16
#> [25,]       0.116837195 1.110223e-16
#> [26,]       0.000000000 1.264212e-01
#> [27,]       0.126036594 0.000000e+00
#> [28,]       0.038228070 0.000000e+00
#> [29,]       0.000000000 1.314567e-01
#> [30,]       0.154080377 1.806984e-01
#> [31,]       0.047444829 1.724705e-01
#> [32,]       0.164707125 2.562013e-02
#> [33,]       0.076180074 2.220446e-16
#> [34,]       0.152047592 1.110223e-16
#> [35,]       0.103148509 1.232429e-01
#> [36,]       0.025992685 7.747127e-02
#> [37,]       0.101012634 0.000000e+00
#> [38,]       0.119952583 2.750924e-02
#> [39,]       0.137792624 0.000000e+00
#> [40,]       0.199780999 2.220446e-16
#> [41,]       0.253154673 1.747075e-01
#> [42,]       0.232809726 0.000000e+00
#> [43,]       0.211197453 0.000000e+00
#> [44,]       0.000000000 5.583468e-02
#> [45,]       0.108771806 9.796969e-02
#> [46,]       0.333658898 0.000000e+00
#> [47,]       0.268515169 4.220483e-02
#> [48,]       0.224431763 7.018923e-02
#> [49,]       0.320117696 1.110223e-16
#> [50,]       0.404552060 7.806218e-02
#> [51,]       0.479062150 9.864966e-02
#> [52,]       0.194584491 2.220446e-16
#> [53,]       0.384389760 0.000000e+00
#> [54,]       0.204414540 0.000000e+00
#> [55,]       0.165219637 7.944489e-02
#> [56,]       0.220109558 6.785312e-02
#> [57,]       0.326009218 2.220446e-16
#> [58,]       0.264994383 1.130324e-02
#> [59,]       0.338849057 2.220446e-16
#> [60,]       0.382278822 0.000000e+00
#> [61,]       0.304204955 8.709801e-02
#> [62,]       0.156971121 1.110223e-16
#> [63,]       0.296492797 9.308505e-04
#> [64,]       0.137975115 0.000000e+00
#> [65,]       0.231373287 2.220446e-16
#> [66,]       0.119567611 0.000000e+00
#> [67,]       0.150320183 2.438535e-02
#> [68,]       0.455844391 6.542046e-02
#> [69,]       0.088836163 1.298467e-01
#> [70,]       0.343819386 0.000000e+00
#> [71,]       0.148741656 0.000000e+00
#> [72,]       0.215744220 9.185609e-02
#> [73,]       0.105329089 2.220446e-16
#> [74,]       0.264289587 2.035537e-01
#> [75,]       0.195868745 1.244088e-02
#> [76,]       0.355075749 0.000000e+00
#> [77,]       0.226123480 0.000000e+00
#> [78,]       0.249759042 8.169634e-02
#> [79,]       0.281987493 1.715041e-01