Synthetic Convergent Dataset

This dataset is a synthetic dataset generated to simulate mixture data for testing and validation purposes in the mixmustr package. It is anchored to empirical values of stable isotopes and fatty acids for a range of plant carbon sources in marine soils stable isotope data (bcs_si), fatty acid data (bcs_fa), and synthetic mixture proportions. synthetic_df_divergent exhibits great difference in the underlying mixing proportions between data streams 1 and 2.

synthetic_df_divergent

Format

A list containing three data frames:

df_stream_1

A data frame containing the simulated mixture data for the first stream, including tracer estimates calculated from the stable isotope and fatty acid data.

df_stream_2

A data frame containing the simulated synthetic proportions for the second stream, a column for each source.

stream_1_props

A data frame containing the synthetic proportions for the first stream, used as input for generating the mixture data. This is only used for testing purposes

Source

The synthetic dataset was generated programmatically using the mixmustr package. The input data sources are:

Stable isotope data: See bcs_si documentation.

Fatty acid data: See bcs_fa documentation.

Details

The dataset was generated using the non-exported make_mixture_data function, which combines stable isotope data (bcs_si), fatty acid data (bcs_fa), and synthetic proportions (stream_1_props and stream_2_props) to produce simulated mixture data. The truth_stream parameter determines which stream's proportions are used as the "true" source contributions.

See also

Examples

data(synthetic_df_divergent)
str(synthetic_df_divergent)
#> List of 3
#>  $ df_stream_1   :'data.frame':	79 obs. of  9 variables:
#>   ..$ group     : chr [1:79] "G1" "G1" "G1" "G1" ...
#>   ..$ 18:1w9    : num [1:79] 11.5 13.1 10.4 12.4 13.4 ...
#>   ..$ 18:2w6    : num [1:79] 14.42 11.26 14.22 9.89 10.37 ...
#>   ..$ 18:3w3    : num [1:79] 22.3 17.6 22.8 13.2 16.7 ...
#>   ..$ 20:4w6    : num [1:79] 0.463 1.041 1.165 0.754 0.578 ...
#>   ..$ 20:5w3    : num [1:79] 0.921 1.628 1.29 1.753 1.574 ...
#>   ..$ 24:0      : num [1:79] 2.23 1.83 1.96 1.9 1.84 ...
#>   ..$ d(13C/12C): num [1:79] -23.7 -22.3 -22.2 -19.8 -21.9 ...
#>   ..$ d(15N/14N): num [1:79] 3.86 4.27 3.57 4.4 4.36 ...
#>  $ df_stream_2   :'data.frame':	79 obs. of  7 variables:
#>   ..$ group            : chr [1:79] "G1" "G1" "G1" "G1" ...
#>   ..$ Macroalgae       : num [1:79] 0 0.0528 0.0385 0 0.0852 ...
#>   ..$ Mangrove         : num [1:79] 0.175 0.485 0.4 0.312 0.331 ...
#>   ..$ Saltmarsh        : num [1:79] 0.1603 0.0617 0.1599 0.3096 0.286 ...
#>   ..$ Seagrass         : num [1:79] 0.0538 0 0 0.0356 0.0242 ...
#>   ..$ SPOM             : num [1:79] 0.344 0.228 0.286 0.166 0.204 ...
#>   ..$ Terrestrial grass: num [1:79] 0.1986 0.1083 0.0495 0.1375 0.069 ...
#>  $ stream_1_props:'data.frame':	79 obs. of  7 variables:
#>   ..$ group            : chr [1:79] "G1" "G1" "G1" "G1" ...
#>   ..$ Macroalgae       : num [1:79] 0 0.0483 0.051 0.0242 0.0173 ...
#>   ..$ Mangrove         : num [1:79] 0.3856 0.292 0.2448 0.0183 0.2622 ...
#>   ..$ Saltmarsh        : num [1:79] 0.347 0.201 0.33 0.456 0.239 ...
#>   ..$ Seagrass         : num [1:79] 0.0177 0.0469 0.063 0.013 0 ...
#>   ..$ SPOM             : num [1:79] 0.183 0.358 0.187 0.42 0.397 ...
#>   ..$ Terrestrial grass: num [1:79] 0.0508 0.0537 0.1239 0 0.0721 ...